requirements.txt
aiogram
aiosqlite
alembic
python-dotenv
SQLAlchemy